+++ src/main/java/com/sch/rezero/dto/community/communityPost/CommunityPostCreate.java
@@ -2,7 +2,6 @@ package com.sch.rezero.dto.community.communityPost;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
-import java.util.List;
 
 public record CommunityPostCreate(
     @Size(max = 255)
