+++ src/main/java/com/sch/rezero/dto/community/communityPost/CommunityUpdateCreate.java
@@ -1,8 +1,6 @@
 package com.sch.rezero.dto.community.communityPost;
 
-import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
-import java.util.List;
 
 public record CommunityUpdateCreate(
     @Size(max = 255)
