+++ src/main/java/com/sch/rezero/dto/recycling/recyclingPost/RecyclingPostUpdateRequest.java
@@ -1,10 +1,6 @@
 package com.sch.rezero.dto.recycling.recyclingPost;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
-
-import java.util.List;
 
 public record RecyclingPostUpdateRequest(
         @Size(max = 255)
@@ -13,9 +9,6 @@ public record RecyclingPostUpdateRequest(
 
         @Size(max = 255)
         String thumbNailImage,
-
-        List<@NotNull Long> imageIds,
-        List<@NotBlank String> imageUrls,
 
         @Size(max = 20)
         String category
