+++ src/main/java/com/sch/rezero/dto/user/profile/UserResponse.java
@@ -6,8 +6,8 @@ public record UserResponse(
     String role,
     String profileUrl,
 
-    Integer follwerCount,
-    Integer follwingCount
+    Integer followerCount,
+    Integer followingCount
 ) {
 
 }
