+++ src/main/java/com/sch/rezero/entity/community/CommunityComment.java
@@ -15,6 +15,8 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
 
 @Entity
 @Getter
@@ -27,11 +29,11 @@ public class CommunityComment {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
-  @ManyToOne(fetch = FetchType.LAZY)
+  @ManyToOne(optional = false, fetch = FetchType.LAZY)
   @JoinColumn(name = "user_id", nullable = false)
   private User user;
 
-  @ManyToOne(fetch = FetchType.LAZY)
+  @ManyToOne(optional = false, fetch = FetchType.LAZY)
   @JoinColumn(name = "community_id", nullable = false)
   private CommunityPost communityPost;
 
@@ -42,9 +44,11 @@ public class CommunityComment {
   @Column(nullable = false)
   private String content;
 
-  @Column(name = "created_at", nullable = false)
+  @CreationTimestamp
+  @Column(name = "created_at", nullable = false, updatable = false)
   private LocalDateTime createdAt;
 
+  @UpdateTimestamp
   @Column(name = "updated_at")
   private LocalDateTime updatedAt;
 
@@ -52,6 +56,5 @@ public class CommunityComment {
     if (content != null && !content.equals(this.content)) {
       this.content = content;
     }
-    this.updatedAt = LocalDateTime.now();
   }
 }
