+++ src/main/java/com/sch/rezero/entity/recycling/Category.java
@@ -16,7 +16,7 @@ public class Category {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, length = 20)
+    @Column(nullable = false)
     private String category;
 
     public void update(String category) {
