+++ src/main/java/com/sch/rezero/entity/user/User.java
@@ -1,6 +1,13 @@
 package com.sch.rezero.entity.user;
 
+import com.sch.rezero.entity.community.CommunityComment;
+import com.sch.rezero.entity.community.CommunityPost;
+import com.sch.rezero.entity.community.Like;
+import com.sch.rezero.entity.recycling.RecyclingPost;
+import com.sch.rezero.entity.recycling.Scrap;
 import jakarta.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -14,19 +21,27 @@ import java.time.LocalDate;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 public class User {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "login_id", nullable = false, length = 50)
+    @Column(name = "login_id", nullable = false)
     private String loginId;
 
     @Column(nullable = false)
     private String password;
 
-    @Column(nullable = false, length = 100)
+    @Column(nullable = false)
     private String name;
 
     private LocalDate birth;
     private String region;
+
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<CommunityPost> communityPosts = new ArrayList<>();
+
+    @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
+    private List<CommunityComment> comments = new ArrayList<>();
+
 }
